requirements.txt
Django==5.1.4
djangorestframework==3.15.2
httpx==0.28.1
itemadapter==0.10.0
reactpy==1.1.0
Requests==2.32.5
Scrapy==2.12.0
selenium==4.39.0
SQLAlchemy==2.0.45
django-cors-headers
psycopg2-binary
dj-database-url
python-decouple